feat: change error handling structure and rename folders
