+++ src/main/java/com/develcode/develfood/exception/ValidationException.java
@@ -1,0 +1,7 @@
+package com.develcode.develfood.exception;
+
+public class ValidationException extends RuntimeException{
+    public ValidationException(String message) {
+        super(message);
+    }
+}
