+++ src/main/java/com/develcode/develfood/services/UserService.java
@@ -11,7 +11,7 @@ import com.develcode.develfood.model.enums.Role;
 import com.develcode.develfood.repository.CustomerRepository;
 import com.develcode.develfood.repository.RestaurantRepository;
 import com.develcode.develfood.repository.UserRepository;
-import com.develcode.develfood.services.exception.validation.EmailExistsException;
+import com.develcode.develfood.services.exception.EmailExistsException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
