+++ src/main/java/com/develcode/develfood/services/exception/CnpjExistsException.java
@@ -1,0 +1,7 @@
+package com.develcode.develfood.services.exception;
+
+import com.develcode.develfood.exception.ValidationException;
+
+public class CnpjExistsException extends ValidationException {
+    public CnpjExistsException(String message) {super(message);}
+}
