+++ src/main/java/com/develcode/develfood/services/exception/EmailExistsException.java
@@ -1,0 +1,7 @@
+package com.develcode.develfood.services.exception;
+
+import jakarta.validation.ValidationException;
+
+public class EmailExistsException extends ValidationException {
+    public EmailExistsException(String message) {super(message);}
+}
