--- src/main/java/com/develcode/develfood/services/exception/validation/CpfExistsException.java
@@ -1,5 +1,0 @@
-package com.develcode.develfood.services.exception.validation;
-
-public class CpfExistsException extends RuntimeException{
-    public CpfExistsException(String message) {super(message);}
-}
