--- src/main/java/com/develcode/develfood/services/exception/validation/EmailExistsException.java
@@ -1,5 +1,0 @@
-package com.develcode.develfood.services.exception.validation;
-
-public class EmailExistsException extends RuntimeException{
-    public EmailExistsException(String message) {super(message);}
-}
