--- src/main/java/com/develcode/develfood/services/exception/validation/ValidationErrorData.java
@@ -1,5 +1,0 @@
-package com.develcode.develfood.services.exception.validation;
-
-public class ValidationErrorData extends RuntimeException{
-
-}
